:+1: Se implementa funcion para determinar el tipo de numero y se reorganizan las dependencias de las bibliotecas.
