+++ tp1/calculator.h
@@ -2,5 +2,6 @@
 #define CALCULATOR_H_INCLUDED
 
 #include <stdio.h>
+#include "input.h"
 
 #endif // CALCULATOR_H_INCLUDED
