+++ tp1/input.h
@@ -5,6 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <float.h>
+#include <math.h>
 
 #define STRING_AS_NUMBER_MAX 64 /**< Buffer de entrada maximo para cadenas numericas */
 #define EXIT_BUFFER '\0' /**< Valor de terminacion de cadena de caracteres. */
