+++ tp1/main.c
@@ -10,7 +10,6 @@
  *
  ********************************************************************/
 #include "menu.h"
-#include "calculator.h"
 
 int main()
 {
