+++ tp1/menu.h
@@ -1,7 +1,7 @@
 #ifndef MENU_H_INCLUDED
 #define MENU_H_INCLUDED
 
-#include "input.h"
+#include "calculator.h"
 
 #define OPTION_MIN 1
 #define OPTION_MAX 9
